tweaked graphics file

+++ Graphics.cpp
@@ -8,13 +8,14 @@ int main()
     const int WINDOW_WIDTH  = 1920;
     const int WINDOW_HEIGHT = 1080;
 
-    sf::RenderWindow window(sf::VideoMode(WINDOW_WIDTH, WINDOW_HEIGHT),
-                            "Scrolling Background with Ship");
+    sf::RenderWindow window(sf::VideoMode(WINDOW_WIDTH, WINDOW_HEIGHT),"Scrolling Background with Ship");
     window.setFramerateLimit(60);
 
     // =====================================================
-    // LOAD TEXTURES
+    // texture loading  --------------------------------------------------------------------------------------
     // =====================================================
+
+    // error Checking
     sf::Texture bgTexture;
     if (!bgTexture.loadFromFile("bg.png"))     // 9600 x 1080
     {
@@ -57,6 +58,7 @@ int main()
         return 1;
     }
 
+    //pixel smoother
     bgTexture.setSmooth(true);
     shipTexture.setSmooth(true);
     wavesTexture.setSmooth(true);
@@ -65,8 +67,9 @@ int main()
     yellowBirdTexture.setSmooth(true);
 
     // =====================================================
-    // BACKGROUND MUSIC
+    // bg music -----------------------------------------------------------------------------------
     // =====================================================
+    
     sf::Music music;
     if (!music.openFromFile("Angry Birds Theme Song.mp3"))
     {
@@ -79,88 +82,93 @@ int main()
     }
 
     // =====================================================
-    // BACKGROUND SETUP (SCROLLING VIEW ON 9600x1080)
+    // bg scroll setup ---------------------------------------------------------------------------
     // =====================================================
+    
     sf::Sprite bgSprite(bgTexture);
 
-    // This rectangle defines the visible 1920x1080 "window"
+    // this rectangle defines the visible 1920x1080 window
     sf::IntRect bgRect(0, 0, WINDOW_WIDTH, WINDOW_HEIGHT);
     bgSprite.setTextureRect(bgRect);
 
     // no scaling now – use 1:1 pixels
-    // bgSprite.setScale(1.0f, 1.0f);
+    // bgSprite.setScale(1.0, 1.0);
 
-    const float BG_SPEED = 200.0f;
+    const float BG_SPEED = 200.0;
 
     // maximum we can shift to the right inside the 9600 texture
-    const int   BG_MAX_OFFSET = static_cast<int>(bgTexture.getSize().x) - WINDOW_WIDTH; // 9600 - 1920 = 7680
-    float       bgOffsetX     = 0.0f;
+    const int BG_MAX_OFFSET = bgTexture.getSize().x - WINDOW_WIDTH;
+    float   bgOffsetX     = 0.0;
 
     // =====================================================
-    // SHIP SETUP
+    // ship -------------------------------------------------------------------------------------------
     // =====================================================
+    
     sf::Sprite shipSprite(shipTexture);
 
-    float desiredShipHeight = WINDOW_HEIGHT * 0.3f;
-    float shipScale         = desiredShipHeight / static_cast<float>(shipTexture.getSize().y);
+    float desiredShipHeight = WINDOW_HEIGHT * 0.3;
+    float shipScale = desiredShipHeight / (shipTexture.getSize().y);
     shipSprite.setScale(shipScale, shipScale);
 
-    float shipX = WINDOW_WIDTH  * 0.08f;
-    float shipY = WINDOW_HEIGHT * 0.65f;
+    float shipX = WINDOW_WIDTH  * 0.08;
+    float shipY = WINDOW_HEIGHT * 0.65;
     shipSprite.setPosition(shipX, shipY);
 
     sf::FloatRect shipBounds = shipSprite.getGlobalBounds();
 
-    float birdDeckX = shipBounds.left + shipBounds.width  * 0.45f;
-    float birdDeckY = shipBounds.top  + shipBounds.height * 0.25f;
+    float birdDeckX = shipBounds.left + shipBounds.width  * 0.45;
+    float birdDeckY = shipBounds.top  + shipBounds.height * 0.25;
 
     // =====================================================
-    // WAVES SETUP (BOTTOM, LOOPING)
+    // wave setup loop ---------------------------------------------------------------------------------
     // =====================================================
+    
     sf::Sprite waves1(wavesTexture);
     sf::Sprite waves2(wavesTexture);
 
     // original texture size
-    float wavesWidth  = static_cast<float>(wavesTexture.getSize().x);
-    float wavesHeight = static_cast<float>(wavesTexture.getSize().y);
+    float wavesWidth  = (wavesTexture.getSize().x);
+    float wavesHeight = (wavesTexture.getSize().y);
 
     // make waves bigger
-    float wavesScaleX = 2.0f;   // 2x wider
-    float wavesScaleY = 1.5f;   // 1.5x taller
+    float wavesScaleX = 2.0;   // scale this is rate 2x
+    float wavesScaleY = 1.5;   
 
     waves1.setScale(wavesScaleX, wavesScaleY);
     waves2.setScale(wavesScaleX, wavesScaleY);
 
-    // size AFTER scaling
+    // size after scaling
     float wavesWidthScaled  = wavesWidth  * wavesScaleX;
     float wavesHeightScaled = wavesHeight * wavesScaleY;
 
     // place so bottom of scaled waves touches bottom of screen
     float wavesY = WINDOW_HEIGHT - wavesHeightScaled;
 
-    waves1.setPosition(0.0f,            wavesY);
+    waves1.setPosition(0.0, wavesY);
     waves2.setPosition(wavesWidthScaled, wavesY);
 
-    const float WAVES_SPEED = 300.0f;
+    const float WAVES_SPEED = 300.0;
 
 
     // =====================================================
-    // BIRDS SETUP
+    // birds setup --------------------------------------------------------------------------------------
     // =====================================================
-    sf::Sprite redBirdSprite(redBirdTexture);
+    
+    sf:: Sprite redBirdSprite(redBirdTexture);
     sf::Sprite blueBirdSprite(blueBirdTexture);
     sf::Sprite yellowBirdSprite(yellowBirdTexture);
 
-    float desiredBirdHeight = shipBounds.height * 0.25f;
+    float desiredBirdHeight = shipBounds.height * 0.25;
 
-    float redScale    = desiredBirdHeight / static_cast<float>(redBirdTexture.getSize().y);
-    float blueScale   = desiredBirdHeight / static_cast<float>(blueBirdTexture.getSize().y);
-    float yellowScale = desiredBirdHeight / static_cast<float>(yellowBirdTexture.getSize().y);
+    float redScale  = desiredBirdHeight / (redBirdTexture.getSize().y);
+    float blueScale = desiredBirdHeight / (blueBirdTexture.getSize().y);
+    float yellowScale = desiredBirdHeight / (yellowBirdTexture.getSize().y);
 
     redBirdSprite.setScale(redScale, redScale);
     blueBirdSprite.setScale(blueScale, blueScale);
     yellowBirdSprite.setScale(yellowScale, yellowScale);
 
+    //keyboard shortcuts
     const int NO_BIRD     = 0;
     const int RED_BIRD    = 1;
     const int BLUE_BIRD   = 2;
@@ -169,16 +177,16 @@ int main()
     int currentBird = NO_BIRD;
 
     // =====================================================
-    // TIME & SCROLL STATE
+    // time, scroll and loop ---------------------------------------------------------------------------
     // =====================================================
     sf::Clock clock;
     bool scrolling = true;
 
     while (window.isOpen())
     {
-        // -------------------------------------------------
-        // EVENT LOOP
-        // -------------------------------------------------
+// -------------------------------------------------
+// loop
+// -------------------------------------------------
         sf::Event event;
         while (window.pollEvent(event))
         {
@@ -219,7 +227,7 @@ int main()
         }
 
         // -------------------------------------------------
-        // UPDATE
+        // update ---------------------------------------------------------------------------------
         // -------------------------------------------------
         float dt = clock.restart().asSeconds();
 
@@ -234,12 +242,12 @@ int main()
                 bgOffsetX -= BG_MAX_OFFSET;  // wrap back (or set to 0)
             }
 
-            bgRect.left = static_cast<int>(bgOffsetX);
+            bgRect.left = (bgOffsetX);
             bgSprite.setTextureRect(bgRect);
 
-            // Waves (still using 2 sprites for loop)
-            waves1.move(-WAVES_SPEED * dt, 0.0f);
-            waves2.move(-WAVES_SPEED * dt, 0.0f);
+            // waves (using 2 sprites for loop)
+            waves1.move(-WAVES_SPEED * dt, 0.0);
+            waves2.move(-WAVES_SPEED * dt, 0.0);
 
             if (waves1.getPosition().x <= -wavesWidthScaled)
             {
@@ -267,11 +275,11 @@ int main()
         }
 
         // -------------------------------------------------
-        // DRAW
+        // obj draw ------------------------------------------------------------------------------------
         // -------------------------------------------------
         window.clear(sf::Color::Black);
 
-        // single background sprite using moving textureRect
+        // single bg sprite using moving textureRect
         window.draw(bgSprite);
 
         window.draw(shipSprite);
